requirements.txt
appnope==0.1.0
asn1crypto==0.24.0
astroid==2.1.0
awscli==1.16.96
backcall==0.1.0
bleach==3.0.2
boto==2.49.0
botocore==1.12.86
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
click==6.7
cloudflare==2.1.0
colorama==0.3.9
configobj==5.0.6
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
Django==1.11.3
django-filter==2.0.0
djangorestframework==3.8.2
docopt==0.6.2
docutils==0.14
ec2==0.4.0
entrypoints==0.3
flake8==3.7.7
future==0.17.1
GoDaddyPy==2.2.7
httpie==1.0.2
idna==2.6
ipykernel==5.1.0
ipyparallel==6.2.3
ipython==7.2.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.4
itchat==1.3.10
itypes==1.1.0
jedi==0.13.2
Jinja2==2.10
jmespath==0.9.3
jsonlines==1.2.0
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.4.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
lazy-object-proxy==1.3.1
lxml==4.3.0
Markdown==3.0.1
MarkupSafe==1.1.0
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.4.0
nbformat==4.4.0
notebook==5.7.4
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prometheus-client==0.5.0
prompt-toolkit==2.0.8
ptyprocess==0.6.0
pyasn1==0.4.5
pycodestyle==2.5.0
pycparser==2.19
pycrypto==2.6.1
pyflakes==2.1.1
Pygments==2.3.1
pylint==2.2.2
PyMySQL==0.9.2
pypng==0.0.19
PyQRCode==1.2.1
python-dateutil==2.7.5
python-gitlab==1.7.0
pytz==2018.6
pyvim==2.0.23
pyvmomi==6.7.1
PyYAML==3.13
pyzabbix==0.7.5
pyzmq==17.1.2
qingstor-sdk==2.2.6
qtconsole==4.4.3
requests==2.21.0
route53==1.0.1
rsa==3.4.2
s3transfer==0.1.13
saws==0.4.3
Send2Trash==1.5.0
six==1.11.0
terminado==0.8.1
testpath==0.4.2
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.2.0
uritemplate==3.0.0
urllib3==1.24.1
wcwidth==0.1.7
webencodings==0.5.1
widgetsnbextension==3.4.2
wrapt==1.11.1
